% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inverted_manhattan_pair.R
\name{generate_plot_title}
\alias{generate_plot_title}
\title{Helper function to generate plot titles with lambda symbol}
\usage{
generate_plot_title(
  plot_title_prefix,
  n_cases,
  n_controls,
  lambda,
  total_snps_in_study,
  add_date_to_title,
  gwas_data,
  font_family = "Arial Unicode MS"
)
}
\arguments{
\item{plot_title_prefix}{Prefix for the plot title}

\item{n_cases}{Number of cases}

\item{n_controls}{Number of controls}

\item{lambda}{Genomic inflation factor}

\item{total_snps_in_study}{Total SNPs in study}

\item{add_date_to_title}{Whether to add date to title}

\item{gwas_data}{GWAS data}

\item{font_family}{Font family for Unicode support}
}
\value{
Formatted plot title string
}
\description{
Helper function to generate plot titles with lambda symbol
}
