% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inverted_manhattan_pair.R
\name{preprocess_manhattan_data}
\alias{preprocess_manhattan_data}
\title{Helper function to preprocess GWAS data}
\usage{
preprocess_manhattan_data(
  gwas_data,
  chr_col,
  bp_col,
  p_col,
  gene_col,
  group_col,
  plot_pval_threshold,
  all_chrs
)
}
\arguments{
\item{gwas_data}{GWAS data}

\item{chr_col}{Chromosome column name}

\item{bp_col}{Base pair column name}

\item{p_col}{P-value column name}

\item{gene_col}{Gene column name}

\item{group_col}{Group column name}

\item{plot_pval_threshold}{P-value threshold}

\item{all_chrs}{All chromosomes to include}
}
\value{
List with processed plot data and axis data
}
\description{
Helper function to preprocess GWAS data
}
